modularizing reducers into user and playlist reducers and use combine reducer in reducer.js
